chore: move CollectionName
