+++ src/common/telemetry_ops/mod.rs
@@ -6,4 +6,3 @@ pub mod distributed_telemetry;
 pub mod hardware;
 pub mod memory_telemetry;
 pub mod requests_telemetry;
-pub mod telemetry_context;
