--- src/common/telemetry_ops/telemetry_context.rs
@@ -1,2 +1,0 @@
-#[derive(Clone, Debug)]
-pub struct CollectionName(pub String);
